refactor: 리포지토리 DIP 적용
