+++ src/main/java/com/example/daobe/objet/domain/repository/ObjetCallRepository.java
@@ -1,5 +1,6 @@
-package com.example.daobe.objet.infrastructure.redis;
+package com.example.daobe.objet.domain.repository;
 
 public interface ObjetCallRepository {
+
     Long getObjetLength(Long objetId);
 }
